+++ src/minesweeper/engine/index.ts
@@ -1,0 +1,3 @@
+export { placeMines } from "./placeMines";
+export { revealCells } from "./revealCells";
+export { toggleFlag } from "./toggleFlag";
